Move HexagonDialogue to trigger, get items working
