--- Assets/Scenes/DialogueTest/HexagonDialogue.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9cd58adddda7996c6921717eb14101d9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
